[Lab2] fix script

+++ Lab2/testall.sh
@@ -1,15 +1,21 @@
 #!/bin/bash
 result=Benchmarks
 
+function prepend() { while read -r line; do echo "${1}${line}"; done; }
+
 route() {
   in=$1
   out=$1.out
   time=$1.txt
-  /usr/bin/time -lp ./router "$in" "$out" 2>&1 | tee "$time"
-}
-test() {
-  time perl ./eval2008.pl "$in" "$out" 2>&1 | tee -a "$result"
-  rm "$out"
+  /usr/bin/time -p ./router "$in" "$out" $((1)) 2>&1 | tee "$time" | prepend "$(basename "$in") "
+  if [ $? ]; then
+    echo "$1 done"
+    time perl ./eval2008.pl "$in" "$out" 2>&1 | tee -a "$result"
+    tail "$time" | tee -a "$result" > /dev/null
+    rm "$out"
+  else
+    echo "route fail"
+  fi
 }
 
 make -j || exit
@@ -17,15 +23,6 @@ rm "$result"
 touch "$result"
 for f in './ISPD 2008 Benchmarks'/*.gr; do
   echo "$f"
-  route "$f"
-  if [ $? ]; then
-    if [ "$1" ]; then
-      test "$f"
-    else
-      rm "$out"
-    fi
-  else
-    echo "route fail"
-  fi
+  route "$f" &
 done
 wait
